Extract app runtime and publication ownership
