+++ backend/app/routes/apps.py
@@ -1,12 +1,9 @@
 """Routes for managing the mini-app registry."""
 
 import asyncio
-import hashlib
-import html
 import io
 import json
 import logging
-import mimetypes
 import os
 import re
 import secrets
@@ -16,12 +13,11 @@ import tempfile
 import time
 import uuid
 from datetime import datetime, UTC
-from pathlib import Path, PurePosixPath
+from pathlib import Path
 from typing import Literal
-from urllib.parse import quote
 
 from fastapi import APIRouter, Depends, HTTPException, Request, Response
-from fastapi.responses import FileResponse, HTMLResponse, StreamingResponse
+from fastapi.responses import StreamingResponse
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session, defer
 
@@ -30,62 +26,38 @@ from app import (
   app_recency, fs_locks, icon_cache, icon_ownership,
   legacy_platform_apps,
   models, providers, schemas,
-  source_dirs, theme,
+  source_dirs,
 )
-from app.artifact_data import (
-  ArtifactDataError,
-  MAX_ARTIFACT_KEYS,
-  MAX_ARTIFACT_TOTAL_BYTES,
-  MAX_ARTIFACT_VALUE_BYTES,
-  artifact_dir_path,
-  artifact_file_path,
-  artifact_usage,
-  list_artifact_keys,
-  canonical_json,
-  parse_json,
-  read_json_file,
-  validate_artifact_id,
-  validate_artifact_key,
+from app.app_identity import (
+  reject_if_source_dir_taken as _reject_if_source_dir_taken,
+  slugify_for_source_dir as _slugify_for_source_dir,
+  validate_source_dir as _validate_source_dir,
 )
-from app.publication import (
-  InvalidPublicationRegistry,
-  PublicationRecord,
-  PublicationReservationConflict,
-  _PUBLISH_PROJECT_RE,
-  _TOKEN_RE,
-  atomic_promote_directory,
-  create_publication_record,
-  new_publication_record,
-  published_root,
-  read_publication_record,
-  registry_path,
-  registry_root,
-  replace_publication_record,
+from app.routes.app_publication import (
+  _revoke_app_publish_tokens,
+  router as publication_router,
 )
-from app import storage_io
+from app.routes.app_runtime import router as runtime_router
 from app.storage_io import (
-  app_dir_usage,
-  atomic_write,
   delete_content_type_tree,
   read_capped_body,
+  rmtree_strict as _rmtree_strict,
 )
 from app.app_capabilities import diff_contracts
 from app.broadcast import get_system_broadcast
 from app.compiler import (
   app_bundle_uses_current_compile_contract,
-  compile_jsx, CompileError,
+  CompileError,
   recompile_app_bundle,
 )
 from app.config import get_settings
 from app.database import get_db
 from app.deps import (
   get_current_owner, get_current_owner_or_app, get_principal, Principal,
-  get_owner_or_app_with_manage_apps, reject_cross_site, resolve_owner_or_app,
+  get_owner_or_app_with_manage_apps, reject_cross_site,
 )
-from app.http_caching import strip_range
 from app.manifest_contract import ManifestContractError, validate_cron_expr
 from app.resource_access import live_app, live_app_or_404
-from app.routes.storage import _recheck_app_identity
 from app.timeutil import now_naive_utc, SOFT_DELETE_TTL
 
 router = APIRouter(prefix="/api/apps", tags=["apps"])
@@ -98,101 +70,7 @@ APP_SOFT_DELETE_TTL = SOFT_DELETE_TTL
 
 log = logging.getLogger("mobius.apps")
 
-def _slugify_for_source_dir(name: str) -> str:
-  """Historical source/storage slug shape used by legacy-row backfill.
-  Lowercase, alphanum + hyphen, collapsed runs, stripped."""
-  slug = "".join(
-    ch if ch.isalnum() else "-" for ch in (name or "").lower()
-  ).strip("-")
-  while "--" in slug:
-    slug = slug.replace("--", "-")
-  slug = slug or "app"
-  # A purely-numeric slug would collide with the numeric-id storage tree:
-  # an app named "123" derives source dir /data/apps/123, which is exactly
-  # where /api/storage/apps/123/... writes land for app id 123. Prefix it
-  # so a source-dir name is never a bare integer.
-  if slug.isdigit():
-    slug = f"app-{slug}"
-  return slug
 
-
-def _derive_source_dir(data_dir: str, name: str) -> str:
-  """Derive `/data/apps/<slug>/` for a legacy row missing source identity."""
-  return str(Path(data_dir) / "apps" / _slugify_for_source_dir(name))
-
-
-def _validate_source_dir(source_dir: str, data_dir: str) -> str:
-  """Validates a caller-supplied source_dir, returning its resolved path.
-
-  App source must be an IMMEDIATE non-numeric child of /data/apps. Everything
-  else is rejected so source_dir cannot point job runners, compilers, or
-  uninstall cleanup at arbitrary paths.
-
-  Raises 400 on either violation. `.resolve()` collapses symlinks and `..`
-  before the containment check.
-  """
-  # resolve() can raise on a pathological path (e.g. a symlink loop). Surface
-  # that as a clean 400, not a 500 (Codex review round-7 #3 robustness caveat).
-  try:
-    resolved = Path(source_dir).resolve()
-  except (OSError, RuntimeError):
-    raise HTTPException(status_code=400, detail="Invalid source_dir.")
-
-  kind = source_dirs.source_dir_kind(resolved, data_dir)
-  if kind == "app":
-    return str(resolved)
-  apps_root = source_dirs.apps_root(data_dir)
-  core_root = source_dirs.platform_core_root(data_dir)
-  if resolved.parent == apps_root and resolved.name.isdigit():
-    raise HTTPException(
-      status_code=400,
-      detail=(
-        "source_dir basename must not be purely numeric — bare integers "
-        "are reserved for the per-app storage path /data/apps/<id>."
-      ),
-    )
-  if resolved.parent == core_root:
-    raise HTTPException(
-      status_code=400,
-      detail="platform core source_dir is no longer an app source root.",
-    )
-  raise HTTPException(
-    status_code=400,
-    detail=(
-      "source_dir must be an immediate non-numeric child of /data/apps."
-    ),
-  )
-
-
-def _reject_if_source_dir_taken(
-  db: Session, source_dir: str, exclude_id: int | None
-) -> None:
-  """Reject (409) if another app already claims this source dir.
-
-  The caller holds ``fs_locks.source_dir_lock(source_dir)``, so the check +
-  the subsequent assignment are atomic against a concurrent first apply.
-  Two apps sharing one source tree makes explicit application ambiguous and
-  uninstall cleanup conservative (it must refuse to rmtree a shared dir), so
-  forbid the duplicate at assignment time. Compared
-  on RESOLVED paths so a symlinked/relative spelling can't smuggle a duplicate.
-  """
-  try:
-    resolved = Path(source_dir).resolve()
-  except (OSError, RuntimeError):
-    return  # a pathological path is rejected by _validate_source_dir already
-  query = db.query(models.App).filter(models.App.source_dir.isnot(None))
-  if exclude_id is not None:
-    query = query.filter(models.App.id != exclude_id)
-  for other in query.all():
-    try:
-      other_resolved = Path(other.source_dir).resolve()
-    except (OSError, RuntimeError):
-      continue
-    if other_resolved == resolved:
-      raise HTTPException(
-        status_code=409,
-        detail="source_dir is already used by another app.",
-      )
 
 
 def _safe_to_rmtree_source(
@@ -346,32 +224,6 @@ def _resolve_app_source_dir(app_source_dir, app_name, settings) -> Path | None:
   return None
 
 
-def _rmtree_strict(path: Path) -> None:
-  """Remove a directory tree, raising if anything survives the attempt.
-
-  Unlike ``shutil.rmtree(ignore_errors=True)``, a failure is surfaced so a
-  caller that is about to free a reusable id or report a wipe succeeded can
-  refuse to do so while data remains on disk. ``rmtree`` raises on the first
-  error and refuses a symlinked root outright, so a symlinked root is handled
-  separately below; the residual ``lexists`` check is the belt-and-suspenders
-  guard for a partial delete that somehow returned without raising.
-  """
-  if path.is_symlink():
-    # A symlinked root must be removed, not followed: shutil.rmtree refuses a
-    # symlink, and Path.exists() follows it (so a DANGLING link would look
-    # already-gone and survive silently — later readable if its target
-    # reappears under a reused id). Unlink the link itself and confirm.
-    path.unlink()
-    if os.path.lexists(path):
-      raise OSError(f"failed to remove symlink {path}")
-    return
-  if not path.exists():
-    return
-  shutil.rmtree(path)
-  if os.path.lexists(path):
-    raise OSError(f"failed to remove {path}")
-
-
 async def _hard_delete_app(db: Session, app: models.App) -> None:
   """Permanently remove an app's DB row, compiled bundle, source tree, and
   id-keyed storage tree — the pre-110 destructive uninstall, now reached only by
@@ -452,49 +304,6 @@ async def _hard_delete_app(db: Session, app: models.App) -> None:
         "Hard-deleted app %s but could not remove its retired source tree",
         deleted_app_id,
       )
-
-
-def allocate_unique_slug(db: Session, name: str, exclude_id: int | None = None) -> str:
-  """Returns a slug that isn't taken by any other App row.
-
-  Starts from the name's slug; if it collides, appends -2, -3, ...
-  until a free one is found. `exclude_id` lets callers re-allocate
-  for an existing row without colliding with itself (e.g. backfill).
-  Slugs pin standalone-install identity (manifest `id`) — keep them
-  stable across renames so home-screen icons don't orphan.
-
-  Deliberately scans ALL rows including tombstoned (deleted_at IS NOT NULL)
-  ones: a soft-deleted app holds its slug until the TTL purge so a
-  reinstall-reattach (which revives the SAME slug) can't be blocked by a new
-  allocation in the recovery window. Do NOT add a deleted_at filter here — it
-  would break that invariant (feature 110).
-  """
-  base = _slugify_for_source_dir(name)
-  candidate = base
-  suffix = 2
-  while True:
-    q = db.query(models.App).filter(models.App.slug == candidate)
-    if exclude_id is not None:
-      q = q.filter(models.App.id != exclude_id)
-    if q.first() is None:
-      return candidate
-    candidate = f"{base}-{suffix}"
-    suffix += 1
-
-
-def ensure_slug(db: Session, app: models.App) -> str:
-  """Returns the app's slug, populating it on first call for legacy rows.
-
-  Apps created before the slug column existed have NULL slug. Lazy
-  backfill on first standalone-route access keeps the migration
-  pure-additive and avoids guessing slugs we might not be able to
-  validate at migration time (uniqueness needs a transaction).
-  """
-  if app.slug:
-    return app.slug
-  app.slug = allocate_unique_slug(db, app.name, exclude_id=app.id)
-  db.commit()
-  return app.slug
 
 
 def _parse_cron_job_line(line: str) -> tuple[str, str] | None:
@@ -3025,1073 +2834,6 @@ def update_app_schedule(
           "zone_cron": None}
 
 
-def _etag_for_app(app: models.App) -> str | None:
-  """Weak ETag derived from `app.updated_at`. Microsecond precision
-  so two updates within the same wall-clock second produce different
-  validators — second-precision risks the agent shipping a fix and
-  the user's cached browser refusing to revalidate."""
-  if not app.updated_at:
-    return None
-  ts_us = int(app.updated_at.timestamp() * 1_000_000)
-  return f'W/"{ts_us}"'
-
-
-def _not_modified_if_match(
-  request: Request,
-  etag: str,
-  offline: bool = False,
-  response_headers: dict[str, str] | None = None,
-) -> Response | None:
-  """Returns a 304 Response if the request's If-None-Match matches
-  `etag`, else None. The 304 keeps the ETag header so a browser
-  re-validating an existing cache entry can keep its validator, and
-  mirrors the X-Mobius-Offline marker so the 304 carries the same
-  cache metadata as the 200 it stands in for. The SW's
-  appCodeStoreAction policy keys on that header for the gated
-  standalone-navigation cache. Callers whose representation metadata changes
-  independently of the body (notably the frame CSP) pass it through so a 304
-  freshens the cached policy instead of preserving obsolete headers."""
-  match = request.headers.get("if-none-match")
-  if match and etag in [v.strip() for v in match.split(",")]:
-    headers = dict(response_headers or {})
-    headers["ETag"] = etag
-    if offline:
-      headers["X-Mobius-Offline"] = "1"
-    return Response(status_code=304, headers=headers)
-  return None
-
-
-_APP_FRAME_CSP = (
-  "sandbox allow-scripts allow-forms allow-popups "
-  "allow-popups-to-escape-sandbox "
-  "allow-top-navigation-by-user-activation"
-)
-
-
-def _frame_etag(
-  app: models.App,
-  frame_path: Path,
-  frame_rev: str | None = None,
-) -> str | None:
-  """Validator for the `/frame` response, combining the app's
-  `updated_at` with the shared runtime-frame file's content and the
-  active theme.
-
-  Unlike the per-app module, the frame serves `app-frame.html` — the
-  isolation boundary + runtime bootstrap — which changes INDEPENDENTLY of any app
-  row. Keying only on `app.updated_at` (as `_etag_for_app` does) means
-  an edit to the frame (e.g. changing the broker protocol) never
-  invalidates an already-installed PWA: it keeps revalidating against
-  an unchanged validator, gets a 304, and runs the stale frame forever.
-  That is exactly how a dropped `/vendor/three/` path pinned clients to
-  a spinner. Folding a hash of the frame's CONTENT in busts every app's
-  frame cache on the next load whenever app-frame.html changes.
-
-  Content hash, not mtime: `cp`, bind-mounts, and backup/restore rewrite
-  mtimes independently of content, which risks UNDER-invalidation (a
-  real content change that keeps its mtime) — the precise failure mode
-  here. The frame file is small, so hashing per request is cheap.
-
-  `frame_rev`: the app-frame.html content hash, already computed once by
-  `load_effective_theme` for the same request. Pass it so the frame file
-  isn't hashed a SECOND time here — the theme bundle and this ETag share
-  one read (both resolve the same candidate list, so the hash is identical;
-  see get_frame). When omitted (None), the hash is computed from
-  `frame_path` as before, so standalone callers and the unit tests are
-  unaffected. An empty rev means the frame was unresolvable — no content
-  part, matching the old read-failure fall-through."""
-  parts: list[str] = []
-  if app.updated_at:
-    parts.append(str(int(app.updated_at.timestamp() * 1_000_000)))
-  if frame_rev is None:
-    try:
-      parts.append(hashlib.sha256(frame_path.read_bytes()).hexdigest()[:16])
-    except OSError:
-      pass
-  elif frame_rev:
-    parts.append(frame_rev)
-  if not parts:
-    return None
-  return 'W/"' + "-".join(parts) + '"'
-
-
-@router.api_route("/{app_id}/frame", methods=["GET", "HEAD"])
-def get_frame(
-  app_id: int,
-  request: Request,
-  db: Session = Depends(get_db),
-):
-  """Serves the mini-app runtime frame HTML.
-
-  Token-free as of 2026-04-27: the parent shell injects the auth
-  token and the current theme via `postMessage` after the iframe
-  loads, instead of having them server-templated into the body.
-
-  Cache freshness model: two independent mechanisms COEXIST. The
-  compound `_frame_etag` (folding `app.updated_at` with the shared
-  frame file's content) plus `Cache-Control: no-cache` drives the
-  browser's HTTP-cache revalidation on cold / non-SW paths — the
-  browser sends `If-None-Match` and gets a 304 when nothing changed
-  or a fresh 200 when `updated_at` advanced or the frame file
-  changed. The service worker revalidates frame/module routes against
-  the same ETag via `appCodeHandler` in `sw.js`; that cache is ungated
-  and applies to every installed app.
-  SEPARATELY, `AppCanvas` appends `?v=<app.updated_at>` to the frame
-  URL, which the SW keeps as its offline cache key (it strips only
-  token/_/install, not `v`), so an app edit changes the SW key and
-  forces a fresh load. `v` is purely a client/SW cache-buster — this
-  endpoint never reads it.
-
-  Frame is intentionally public — it's just the runtime shell
-  (error UI, postMessage broker/bootstrap). Actual app
-  modules at `/api/apps/{id}/module` still require a token. An
-  attacker embedding this frame in their own page would receive
-  the iframe's `moebius:frame-mounted` postMessage on their parent window,
-  but the iframe's origin check (against `window.location.origin`)
-  rejects any reply from a non-Möbius origin, so no token can be
-  coerced into the frame.
-  """
-  app = live_app(db, app_id)
-  if not app or not app.compiled_path:
-    raise HTTPException(status_code=404, detail="App not found.")
-  compiled = Path(app.compiled_path)
-  if not compiled.exists():
-    raise HTTPException(status_code=404, detail="Compiled module missing.")
-
-  # Frame priority: served platform frontend first, then the baked-in fallback.
-  # Resolve this BEFORE the ETag so the validator reflects the frame file's
-  # content (see _frame_etag) — otherwise a changed frame never reaches
-  # installed PWAs.
-  frame_candidates = [
-    Path(get_settings().data_dir)
-    / "platform" / "frontend" / "public" / "app-frame.html",
-    # Repo-relative dev/test fallback (== served clone in-container).
-    Path(__file__).resolve().parents[3] / "frontend" / "public" / "app-frame.html",
-    Path("/app/app-frame.html"),
-  ]
-  frame_path = next((p for p in frame_candidates if p.exists()), None)
-  if frame_path is None:
-    raise HTTPException(status_code=404, detail="Frame not found.")
-
-  # The frame is no longer theme-varying: theme-as-data moved theming to the
-  # client (the frame's pre-paint IIFE reads the __mobius-theme__ slot +
-  # localStorage and paints flash-free; the server no longer injects a
-  # <style>). So the validator keys only on app.updated_at + the
-  # app-frame.html content hash — NOT the theme. A light/dark toggle no
-  # longer needs to bust the frame cache, because the served frame bytes
-  # don't change with the theme. Compute the frame content hash and key the
-  # validator on it plus app.updated_at.
-  frame_rev = theme.frame_content_rev(get_settings().data_dir)
-  etag = _frame_etag(app, frame_path, frame_rev=frame_rev)
-  frame_cache_headers = {
-    "Cache-Control": "no-cache",
-    "Content-Security-Policy": _APP_FRAME_CSP,
-  }
-  if etag:
-    not_modified = _not_modified_if_match(
-      request,
-      etag,
-      app.offline_capable,
-      response_headers=frame_cache_headers,
-    )
-    if not_modified is not None:
-      return not_modified
-
-  html = frame_path.read_text(encoding="utf-8")
-
-  # Per-app server-side substitution of the app/chat ids the runtime needs.
-  html = html.replace(
-    "var _FRAME_APP_ID = 'unknown'",
-    f"var _FRAME_APP_ID = {json.dumps(str(app_id))}",
-  )
-  html = html.replace(
-    "var _FRAME_CHAT_ID = ''",
-    f"var _FRAME_CHAT_ID = {json.dumps(app.chat_id or '')}",
-  )
-
-  # Theme-as-data: the frame no longer has the theme server-injected. Its
-  # pre-paint IIFE reads the __mobius-theme__ slot (when the server fills
-  # one) and the shell's same-origin localStorage to paint --bg / data-theme
-  # / color-scheme flash-free from the fallback :root + the persisted owner
-  # mode. The parent shell still posts moebius:frame-init/-theme for LIVE
-  # swaps without a reload. Removing the injection means the served frame
-  # bytes are theme-independent (so the ETag no longer folds the theme).
-
-  # The element remains unsandboxed until navigation so the shell service
-  # worker can intercept and serve a cached frame offline. Apply the equivalent
-  # sandbox on the RESPONSE: the loaded app still receives an opaque origin,
-  # including when this backend is reached without the edge proxy. Caddy adds
-  # the full resource policy while preserving this sandbox contract. Popups
-  # opened by an explicit app link must escape the opaque-origin sandbox:
-  # otherwise the destination inherits Origin: null and sites such as GitHub
-  # load their document but fail same-origin API/storage requests. This does
-  # not relax the app frame itself or let it navigate the owner shell.
-  headers = dict(frame_cache_headers)
-  if etag:
-    headers["ETag"] = etag
-  # The X-Mobius-Offline header does not gate frame/module caching: the SW
-  # caches code for every installed app via appCodeHandler(OFFLINE_APPS_CACHE,
-  # {gated:false}), regardless of this header. It only gates the separate
-  # standalone-navigation cache and offline write/open semantics.
-  # Offline capability is a function of server state, not a client-pushed list.
-  if app.offline_capable:
-    headers["X-Mobius-Offline"] = "1"
-
-  # app_open: emit on the GET 200 path only — the 304 short-circuit above
-  # already returned for cache-revalidating loads (which would otherwise
-  # double-count every freshness check on a navigation back), and a HEAD is
-  # an existence probe, not a real open, so it must not count either. Best-
-  # effort: a log failure must not block the frame response
-  # (activity.log_event swallows its own OSError).
-  if request.method != "HEAD":
-    activity.log_event(
-      "app_open", app_id=app.id, slug=ensure_slug(db, app),
-    )
-  return HTMLResponse(html, headers=headers)
-
-
-@router.api_route("/{app_id}/module", methods=["GET", "HEAD"])
-def get_module(
-  app_id: int,
-  request: Request,
-  token: str | None = None,
-  db: Session = Depends(get_db),
-):
-  """Serves the compiled JS module for a mini-app.
-
-  Accepts a `token` query parameter so the iframe can load the
-  module without custom request headers (dynamic `import()` doesn't
-  set an Authorization header).
-
-  Cache freshness: ETag derived from `app.updated_at` (microsecond
-  precision) + `Cache-Control: no-cache`. Browser sends
-  `If-None-Match` on every fetch; we return 304 when the app hasn't
-  changed. Matches the `/frame` route's strategy — see comment
-  there for the broader rationale.
-  """
-  # Apps share modules same as they share storage — every mini-app
-  # is authored by the owner's own agent, and a multi-app workflow
-  # may legitimately want to import or interop across them. Any
-  # valid token (owner or app-scoped) is allowed to fetch any
-  # module by id. See CLAUDE.md "Mini-app sandbox — accepted
-  # same-origin decision" for the broader trust model. resolve_owner_
-  # or_app runs the same decode + revocation check the header deps use,
-  # so a signed-out token can't keep pulling module source; the empty-
-  # token guard stays explicit to keep the "Valid token required" 401
-  # (and to avoid feeding a None token into the JWT decoder).
-  if not token:
-    raise HTTPException(
-      status_code=401, detail="Valid token required."
-    )
-  resolve_owner_or_app(token, db)
-  app = live_app(db, app_id)
-  if not app or not app.compiled_path:
-    raise HTTPException(status_code=404, detail="Module not found.")
-  path = Path(app.compiled_path)
-  etag = _etag_for_app(app)
-  offline_capable = bool(app.offline_capable)
-  # FileResponse streams after this function returns. Do not make the stream's
-  # lifetime the database checkout's lifetime.
-  db.close()
-  if not path.exists():
-    raise HTTPException(
-      status_code=404, detail="Compiled module not found on disk."
-    )
-
-  if etag:
-    not_modified = _not_modified_if_match(request, etag, offline_capable)
-    if not_modified is not None:
-      return not_modified
-
-  headers = {"Cache-Control": "no-cache"}
-  if etag:
-    headers["ETag"] = etag
-  # See get_frame: X-Mobius-Offline does not gate in-shell module caching.
-  # The SW caches modules for every installed app regardless of this header;
-  # the header only gates the separate standalone-navigation cache and
-  # offline write/open semantics.
-  if offline_capable:
-    headers["X-Mobius-Offline"] = "1"
-  # The module is a REVALIDATING response (no-cache + stable ETag), so it
-  # must never answer a 206. A `Range: bytes=0-0` probe of a FileResponse
-  # would otherwise let Chromium store the 1-byte slice and later serve it
-  # as a status-200 full body — a black mini-app until the next app update.
-  # Stripping Range here keeps the streamed full-body 200 (see http_caching).
-  strip_range(request)
-  return FileResponse(
-    path,
-    media_type="application/javascript",
-    headers=headers,
-  )
-
-
-@router.get("/{app_id}/validate")
-async def validate_app(
-  app_id: int,
-  db: Session = Depends(get_db),
-  _: models.Owner = Depends(get_current_owner),
-):
-  """Validates a compiled mini-app for common issues.
-
-  Checks that the compiled file exists, is parseable JS, exports a
-  default, and that the source JSX is present. Returns a report the
-  agent can use to decide whether to offer debugging.
-  """
-  app = live_app_or_404(db, app_id)
-  app_name = app.name
-  jsx_source = app.jsx_source
-  compiled_path = app.compiled_path
-  db.close()
-
-  issues = []
-
-  if not jsx_source:
-    issues.append("No JSX source stored in database.")
-  if not compiled_path:
-    issues.append("No compiled path set — compilation may have failed.")
-  else:
-    path = Path(compiled_path)
-    if not path.exists():
-      issues.append(
-        f"Compiled file missing at {compiled_path}."
-      )
-    else:
-      js = path.read_text(encoding="utf-8")
-      if not js.strip():
-        issues.append("Compiled file is empty.")
-      elif not re.search(r"export\s+default\b|export\s*\{[^}]*\bas\s+default\b", js):
-        issues.append(
-          "Compiled JS has no default export — "
-          "the component won't mount."
-        )
-      # Quick syntax check via node --check if available. Uses
-      # asyncio.create_subprocess_exec so the FastAPI event loop
-      # stays free while node runs (a blocking subprocess.run here
-      # would stall every other request for up to the 5s timeout).
-      proc = None
-      try:
-        proc = await asyncio.create_subprocess_exec(
-          "node", "--check", str(path),
-          stdout=asyncio.subprocess.PIPE,
-          stderr=asyncio.subprocess.PIPE,
-        )
-        try:
-          stdout_b, stderr_b = await asyncio.wait_for(
-            proc.communicate(), timeout=5,
-          )
-        except asyncio.TimeoutError:
-          # Kill the orphan node process; otherwise it lingers
-          # holding the pipe open until the OS reaps it.
-          try:
-            proc.kill()
-            await proc.wait()
-          except ProcessLookupError:
-            pass
-          issues.append("Syntax check timed out.")
-        else:
-          if proc.returncode != 0:
-            stderr = stderr_b.decode("utf-8", errors="replace")
-            issues.append(
-              f"JS syntax error: {stderr.strip()}"
-            )
-      except FileNotFoundError:
-        pass  # node not available — skip this check
-
-  return {
-    "app_id": app_id,
-    "name": app_name,
-    "valid": len(issues) == 0,
-    "issues": issues,
-  }
-
-
-
-# ---- Artifact persistence + published site ownership ------------------
-
-
-def _read_publish_token_hint(token_file: Path) -> str | None:
-  """Read the app-writable token hint without following its symlink."""
-  if token_file.is_symlink():
-    return None
-  try:
-    info = token_file.stat()
-  except OSError:
-    return None
-  if not info.st_size or info.st_size > 128 or not token_file.is_file():
-    return None
-  flags = os.O_RDONLY
-  if hasattr(os, "O_NOFOLLOW"):
-    flags |= os.O_NOFOLLOW
-  try:
-    fd = os.open(token_file, flags)
-  except OSError:
-    return None
-  try:
-    raw = os.read(fd, 129)
-  finally:
-    os.close(fd)
-  try:
-    token = raw.decode("utf-8").strip()
-  except UnicodeDecodeError:
-    return None
-  return token if _TOKEN_RE.fullmatch(token) else None
-
-
-def _registry_records_for_app(settings, app_id: int) -> list[PublicationRecord]:
-  root = registry_root(settings)
-  if root.is_symlink() or not root.is_dir():
-    return []
-  records = []
-  for path in root.glob("*.json"):
-    token = path.stem
-    if not _TOKEN_RE.fullmatch(token):
-      continue
-    try:
-      record = read_publication_record(settings, token)
-    except InvalidPublicationRegistry as exc:
-      log.warning("publish registry %s is invalid: %s", token, exc)
-      continue
-    if record is not None and record.app_id == app_id:
-      records.append(record)
-  return records
-
-
-def _legacy_project_hint(storage: Path, token_file: Path) -> str | None:
-  try:
-    rel = token_file.relative_to(storage)
-  except ValueError:
-    return None
-  if rel.parts == ("build", "publish-token.txt"):
-    return None
-  if (
-    len(rel.parts) == 4
-    and rel.parts[0] == "projects"
-    and rel.parts[2:] == ("build", "publish-token.txt")
-    and _PUBLISH_PROJECT_RE.fullmatch(rel.parts[1])
-  ):
-    return rel.parts[1]
-  return None
-
-
-async def _revoke_publish_token(
-  settings,
-  app_id: int,
-  app_gen: str | None,
-  token: str,
-  project_id: str | None,
-) -> bool:
-  """Permanently revoke one owned token before physical cleanup.
-
-  Returns whether the token is now durably un-servable — either because the
-  revocation was written or because this app never owned it. False means a
-  reservation the caller asked about is STILL ACTIVE, so a caller that reports
-  success to the owner must not ignore it: the page would stay public.
-  """
-  if not _TOKEN_RE.fullmatch(token or ""):
-    return True
-  try:
-    record = read_publication_record(settings, token)
-  except InvalidPublicationRegistry as exc:
-    # A corrupt reservation already fails closed.  Do not let an app-writable
-    # hint authorize deleting the unknown reservation's snapshot.
-    log.warning("cannot revoke invalid publication %s: %s", token, exc)
-    return False
-  if record is None:
-    # No reservation exists, so nothing here proves this app owns the token.
-    # The only thing naming it is publish-token.txt, which lives in app-
-    # writable storage — any app can plant another app's token there and would
-    # otherwise get that app's public snapshot deleted. The registry is the
-    # sole ownership authority: a hint may POINT AT a record, never create one.
-    # Pre-registry snapshots are therefore inert rather than hint-revocable;
-    # removing one is an explicit owner action, not an app-triggered side
-    # effect.
-    log.warning(
-      "ignoring unregistered publish-token hint %s while revoking app %s",
-      token, app_id,
-    )
-    return True
-  if record.app_id != app_id:
-    log.warning(
-      "ignoring publish-token hint %s owned by app %s while revoking app %s",
-      token, record.app_id, app_id,
-    )
-    return True
-  # A hint names a token but does not prove which PROJECT it belongs to, and
-  # hints live in app-writable storage. Checking app_id alone would let one
-  # project's stray hint permanently revoke and delete a sibling project's
-  # publication inside the same app, so honor the hint only when the registry
-  # agrees on the whole binding. `project_id=None` means the caller is tearing
-  # the app down wholesale and every project of the live generation goes.
-  if project_id is not None and record.project_id != project_id:
-    log.warning(
-      "ignoring publish-token hint %s for project %s while revoking project %s",
-      token, record.project_id, project_id,
-    )
-    return True
-  try:
-    if record.state != "revoked":
-      record = replace_publication_record(settings, record, "revoked")
-  except (OSError, InvalidPublicationRegistry,
-          PublicationReservationConflict) as exc:
-    log.error("failed to persist revocation for token %s: %s", token, exc)
-    return False
-
-  # The durable revoked state is written before either best-effort rmtree.
-  for root_name in ("published", "published-data"):
-    root = Path(settings.data_dir) / root_name
-    target = root / token
-    if root.is_symlink() or target.is_symlink():
-      log.error("refusing symlink publication cleanup: %s", target)
-      continue
-    if not target.exists():
-      continue
-    try:
-      await asyncio.to_thread(shutil.rmtree, target)
-    except OSError as exc:
-      log.warning("revoked token %s cleanup failed for %s: %s",
-                  token, target, exc)
-  # The snapshot rmtree above is best-effort cleanup; the durable `revoked`
-  # record already makes the token un-servable, so reaching here is success.
-  return True
-
-
-async def _revoke_app_publish_tokens(
-  settings,
-  app_id: int,
-  app_gen: str | None,
-) -> None:
-  """Revoke registry-owned and legacy tokens while app storage still exists.
-
-  The caller holds ``app_storage_lock(app_id)``.  Each token is independent so
-  one corrupt record or failed rmtree cannot prevent revoking the rest.
-  """
-  tokens: dict[str, str | None] = {
-    record.token: record.project_id
-    for record in _registry_records_for_app(settings, app_id)
-  }
-  storage = Path(settings.data_dir) / "apps" / str(app_id)
-  if not storage.is_symlink() and storage.is_dir():
-    try:
-      token_files = list(storage.rglob("build/publish-token.txt"))
-    except OSError as exc:
-      log.warning("legacy publish-token scan failed for app %s: %s", app_id, exc)
-      token_files = []
-    for token_file in token_files:
-      token = _read_publish_token_hint(token_file)
-      if token is not None:
-        tokens.setdefault(token, _legacy_project_hint(storage, token_file))
-  for token, project_id in tokens.items():
-    try:
-      await _revoke_publish_token(
-        settings, app_id, app_gen, token, project_id,
-      )
-    except Exception as exc:  # best-effort batch boundary
-      log.exception("unexpected publication revoke failure for %s: %s",
-                    token, exc)
-
-
-@router.api_route(
-  "/{app_id}/artifact-data/{artifact_id}",
-  methods=["GET"],
-  dependencies=[Depends(reject_cross_site)],
-)
-async def artifact_data_keys(
-  app_id: int,
-  artifact_id: str,
-  db: Session = Depends(get_db),
-  principal: Principal = Depends(get_principal),
-):
-  """List an artifact's stored keys, derived from the directory.
-
-  The keys are enumerated server-side precisely so no client has to maintain an
-  index file: two tabs updating one would race and silently drop a key. The
-  directory cannot disagree with itself.
-  """
-  if principal.app_id is not None and principal.app_id != app_id:
-    raise HTTPException(403, "An app may only access its own artifact data.")
-  if not validate_artifact_id(artifact_id):
-    raise HTTPException(400, "Invalid artifact_id.")
-  app = live_app_or_404(db, app_id)
-  expected_nonce = app.token_nonce
-  settings = get_settings()
-  async with fs_locks.app_storage_lock(app_id):
-    _recheck_app_identity(db, app_id, expected_nonce)
-    try:
-      keys = list_artifact_keys(
-        artifact_dir_path(settings, app_id, artifact_id),
-      )
-    except ArtifactDataError as exc:
-      raise HTTPException(400, str(exc)) from exc
-  return {"keys": keys}
-
-
-@router.api_route(
-  "/{app_id}/artifact-data/{artifact_id}/{key}",
-  methods=["GET", "PUT", "DELETE"],
-  dependencies=[Depends(reject_cross_site)],
-)
-async def artifact_data_value(
-  app_id: int,
-  artifact_id: str,
-  key: str,
-  request: Request,
-  db: Session = Depends(get_db),
-  principal: Principal = Depends(get_principal),
-):
-  """Read or mutate one server-validated, quota-bound artifact JSON key."""
-  if principal.app_id is not None and principal.app_id != app_id:
-    raise HTTPException(403, "An app may only access its own artifact data.")
-  if not validate_artifact_id(artifact_id) or not validate_artifact_key(key):
-    raise HTTPException(400, "Invalid artifact_id or key.")
-  app = live_app_or_404(db, app_id)
-  expected_nonce = app.token_nonce
-  value_bytes = None
-  if request.method == "PUT":
-    raw = await read_capped_body(request, cap=MAX_ARTIFACT_VALUE_BYTES)
-    try:
-      value_bytes = canonical_json(parse_json(raw))
-    except ArtifactDataError as exc:
-      raise HTTPException(400, str(exc)) from exc
-    if len(value_bytes) > MAX_ARTIFACT_VALUE_BYTES:
-      raise HTTPException(413, "Artifact value exceeds 64 KB.")
-
-  settings = get_settings()
-  async with fs_locks.app_storage_lock(app_id):
-    _recheck_app_identity(db, app_id, expected_nonce)
-    try:
-      artifact_root, file_path = artifact_file_path(
-        settings, app_id, artifact_id, key,
-      )
-    except ArtifactDataError as exc:
-      raise HTTPException(400, str(exc)) from exc
-
-    if request.method == "GET":
-      try:
-        return read_json_file(file_path)
-      except ArtifactDataError as exc:
-        raise HTTPException(404, "Artifact value not found.") from exc
-
-    if request.method == "DELETE":
-      if file_path.is_symlink() or not file_path.is_file():
-        raise HTTPException(404, "Artifact value not found.")
-      file_path.unlink()
-      return Response(status_code=204)
-
-    try:
-      total, key_count = artifact_usage(artifact_root)
-    except ArtifactDataError as exc:
-      raise HTTPException(400, str(exc)) from exc
-    try:
-      old_size = file_path.stat().st_size if file_path.is_file() else 0
-    except OSError:
-      old_size = 0
-    is_new_key = not file_path.is_file()
-    if is_new_key and key_count >= MAX_ARTIFACT_KEYS:
-      raise HTTPException(400, "Artifact data is limited to 100 keys.")
-    projected = total - old_size + len(value_bytes)
-    if projected > MAX_ARTIFACT_TOTAL_BYTES:
-      raise HTTPException(413, "Artifact data exceeds the 1 MB quota.")
-    # The per-artifact caps above bound ONE namespace, and artifact_id is
-    # caller-chosen — inventing namespaces would otherwise multiply them
-    # without limit. The per-app backstop every other storage write already
-    # honors is what actually bounds the tree, so charge this write against it
-    # too. Read the cap from the module so a test can shrink it.
-    app_dir = Path(settings.data_dir) / "apps" / str(app_id)
-    app_projected = app_dir_usage(app_dir) - old_size + len(value_bytes)
-    if app_projected > storage_io.MAX_APP_STORAGE_BYTES:
-      raise HTTPException(
-        413,
-        "App storage quota exceeded — this write would bring the app to "
-        f"{app_projected} bytes, over the "
-        f"{storage_io.MAX_APP_STORAGE_BYTES}-byte per-app limit.",
-      )
-    atomic_write(file_path, value_bytes)
-  return Response(status_code=204)
-
-
-class LinkPreviewRequest(BaseModel):
-  title: str = Field(min_length=1, max_length=200)
-  description: str | None = Field(default=None, max_length=500)
-  image_path: str | None = Field(default=None, max_length=240)
-  image_alt: str | None = Field(default=None, max_length=300)
-  image_width: int | None = Field(default=None, ge=1, le=10000)
-  image_height: int | None = Field(default=None, ge=1, le=10000)
-  site_name: str | None = Field(default=None, max_length=100)
-
-
-class PublishRequest(BaseModel):
-  project_id: str | None = None
-  link_preview: LinkPreviewRequest | None = None
-
-
-def _publication_url(settings, token: str) -> str:
-  origin = settings.frontend_origin.rstrip("/")
-  return f"{origin}/sites/{token}/"
-
-
-def _validated_preview_image(
-  site_dir: Path,
-  image_path: str | None,
-) -> tuple[str, str] | None:
-  if image_path is None:
-    return None
-  raw = image_path.strip()
-  rel = PurePosixPath(raw)
-  if (
-    not raw
-    or rel.is_absolute()
-    or any(part in ("", ".", "..") for part in rel.parts)
-    or "\\" in raw
-  ):
-    raise HTTPException(422, "invalid link preview image_path")
-  target = site_dir.joinpath(*rel.parts)
-  if not target.is_file() or target.is_symlink():
-    raise HTTPException(400, "Link preview image is missing from the site.")
-  content_type, _encoding = mimetypes.guess_type(raw)
-  if content_type not in {
-    "image/gif", "image/jpeg", "image/png", "image/webp",
-  }:
-    raise HTTPException(422, "unsupported link preview image type")
-  return quote(raw, safe="/"), content_type
-
-
-def _link_preview_markup(
-  settings,
-  token: str,
-  preview: LinkPreviewRequest,
-  site_dir: Path,
-) -> bytes:
-  title = preview.title.strip()
-  if not title:
-    raise HTTPException(422, "link preview title cannot be blank")
-  canonical = _publication_url(settings, token)
-  image = _validated_preview_image(site_dir, preview.image_path)
-  tags = [
-    '<meta property="og:type" content="website">',
-    f'<meta property="og:title" content="{html.escape(title, quote=True)}">',
-    f'<meta property="og:url" content="{html.escape(canonical, quote=True)}">',
-    f'<link rel="canonical" href="{html.escape(canonical, quote=True)}">',
-  ]
-  description = (preview.description or "").strip()
-  if description:
-    escaped = html.escape(description, quote=True)
-    tags.append(f'<meta property="og:description" content="{escaped}">')
-  site_name = (preview.site_name or "").strip()
-  if site_name:
-    tags.append(
-      f'<meta property="og:site_name" '
-      f'content="{html.escape(site_name, quote=True)}">',
-    )
-  if image:
-    image_path, content_type = image
-    image_url = f"{canonical}{image_path}"
-    escaped_url = html.escape(image_url, quote=True)
-    tags.extend((
-      f'<meta property="og:image" content="{escaped_url}">',
-      f'<meta property="og:image:type" content="{content_type}">',
-    ))
-    if preview.image_width is not None:
-      tags.append(
-        f'<meta property="og:image:width" content="{preview.image_width}">',
-      )
-    if preview.image_height is not None:
-      tags.append(
-        f'<meta property="og:image:height" content="{preview.image_height}">',
-      )
-    image_alt = (preview.image_alt or "").strip()
-    if image_alt:
-      escaped_alt = html.escape(image_alt, quote=True)
-      tags.append(f'<meta property="og:image:alt" content="{escaped_alt}">')
-    tags.extend((
-      '<meta name="twitter:card" content="summary_large_image">',
-      f'<meta name="twitter:title" '
-      f'content="{html.escape(title, quote=True)}">',
-    ))
-    if description:
-      tags.append(
-        f'<meta name="twitter:description" '
-        f'content="{html.escape(description, quote=True)}">',
-      )
-    tags.append(f'<meta name="twitter:image" content="{escaped_url}">')
-    if image_alt:
-      tags.append(
-        f'<meta name="twitter:image:alt" content="{escaped_alt}">',
-      )
-  block = (
-    "\n<!-- mobius:link-preview:start -->\n"
-    + "\n".join(tags)
-    + "\n<!-- mobius:link-preview:end -->"
-  )
-  return block.encode("utf-8")
-
-
-def _inject_link_preview(
-  settings,
-  token: str,
-  preview: LinkPreviewRequest,
-  site_dir: Path,
-) -> None:
-  index = site_dir / "index.html"
-  if not index.is_file() or index.is_symlink():
-    raise HTTPException(400, "Link previews require a site index.html.")
-  source = index.read_bytes()
-  head = re.search(br"<head(?:\s[^>]*)?>", source, flags=re.IGNORECASE)
-  if head is None:
-    raise HTTPException(400, "Link previews require an HTML head.")
-  markup = _link_preview_markup(settings, token, preview, site_dir)
-  atomic_write(index, source[:head.end()] + markup + source[head.end():])
-
-
-def _publish_paths(settings, app, project_id: str | None):
-  storage = Path(settings.data_dir) / "apps" / str(app.id)
-  base = storage / "projects" / project_id if project_id else storage
-  return base / "build" / "site", base / "build" / "publish-token.txt"
-
-
-def _validate_publish_paths(settings, app, project_id: str | None) -> None:
-  storage = Path(settings.data_dir) / "apps" / str(app.id)
-  site_dir, token_file = _publish_paths(settings, app, project_id)
-  components = [storage]
-  if project_id is not None:
-    components.extend((storage / "projects", storage / "projects" / project_id))
-  components.extend((site_dir.parent, site_dir, token_file))
-  if any(path.is_symlink() for path in components):
-    raise HTTPException(400, "Symlinks are not allowed in publish paths.")
-  storage_resolved = storage.resolve()
-  site_resolved = site_dir.resolve()
-  if storage_resolved not in site_resolved.parents:
-    raise HTTPException(400, "Publish path escaped app storage.")
-
-
-def _mint_publish_record(settings, app, project_id: str | None):
-  while True:
-    token = uuid.uuid4().hex
-    if os.path.lexists(registry_path(settings, token)):
-      continue
-    if os.path.lexists(published_root(settings) / token):
-      continue
-    record = new_publication_record(
-      token, app.id, app.token_nonce, project_id, state="staged",
-    )
-    try:
-      create_publication_record(settings, record)
-    except PublicationReservationConflict:
-      continue
-    return token, record
-
-
-@router.post("/{app_id}/publish", dependencies=[Depends(reject_cross_site)])
-async def publish_app_site(
-  app_id: int,
-  body: PublishRequest,
-  db: Session = Depends(get_db),
-  principal: Principal = Depends(get_principal),
-):
-  """Publish a project's built static site to a stable token URL.
-
-  Snapshots <storage>/[projects/<pid>/]build/site/ to
-  <data_dir>/published/<token>/ and returns /sites/<token>/. The token is
-  stable per project (kept in the project's build/ dir) so re-publishing
-  updates the SAME URL. Owner or the app's own token only.
-  """
-  if principal.app_id is not None and principal.app_id != app_id:
-    raise HTTPException(403, "An app may only publish its own site.")
-  project_id = (body.project_id or "").strip() or None
-  if project_id is not None and not _PUBLISH_PROJECT_RE.match(project_id):
-    raise HTTPException(422, "invalid project_id")
-  app = live_app_or_404(db, app_id)
-  expected_nonce = app.token_nonce
-  settings = get_settings()
-  async with fs_locks.app_storage_lock(app_id):
-    _recheck_app_identity(db, app_id, expected_nonce)
-    _validate_publish_paths(settings, app, project_id)
-    site_dir, token_file = _publish_paths(settings, app, project_id)
-    try:
-      site_ready = site_dir.is_dir() and any(site_dir.iterdir())
-    except OSError:
-      site_ready = False
-    if not site_ready:
-      raise HTTPException(
-        400, "No built site to publish — build the project first.",
-      )
-
-    token = _read_publish_token_hint(token_file)
-    record = None
-    if token is not None:
-      try:
-        hinted = read_publication_record(settings, token)
-      except InvalidPublicationRegistry:
-        hinted = False
-      if (
-        isinstance(hinted, PublicationRecord)
-        and hinted.binding() == (app.id, app.token_nonce, project_id)
-        and hinted.state == "active"
-      ):
-        # Re-publishing an app's OWN registered token keeps its URL stable.
-        record = hinted
-      # An unregistered token is deliberately NOT adopted. publish-token.txt
-      # sits in app-writable storage, so adopting a token merely because a
-      # hint names it and published/<token>/ happens to exist would let any
-      # app claim another app's already-shared public URL and overwrite its
-      # content. The registry is the sole ownership authority, so an
-      # unrecognized hint falls through to minting a fresh token below; the
-      # pre-registry snapshot keeps serving its old content untouched.
-    republishing = record is not None
-    if record is None:
-      token, record = _mint_publish_record(settings, app, project_id)
-
-    root = published_root(settings)
-    staging_root = root / ".staging"
-    if root.is_symlink() or staging_root.is_symlink():
-      raise HTTPException(400, "Invalid published staging directory.")
-    staging_root.mkdir(parents=True, exist_ok=True)
-    stage = staging_root / uuid.uuid4().hex
-    destination = root / token
-    had_destination = destination.exists()
-
-    def _snapshot():
-      if any(path.is_symlink() for path in site_dir.rglob("*")):
-        raise HTTPException(
-          400, "Built site contains symlinks; refusing to publish.",
-        )
-      shutil.copytree(site_dir, stage, symlinks=True)
-      if body.link_preview is not None:
-        _inject_link_preview(
-          settings, token, body.link_preview, stage,
-        )
-
-    promoted = False
-    preserve_stage = False
-    try:
-      await asyncio.to_thread(_snapshot)
-      # Keep the exchange itself on this task so cancellation cannot leave the
-      # thread committing a swap after ``promoted`` incorrectly stayed false.
-      atomic_promote_directory(stage, destination)
-      promoted = True
-      atomic_write(token_file, token)
-      record = replace_publication_record(settings, record, "active")
-    except BaseException as publish_exc:
-      if record.state == "staged":
-        # A first publish has no prior public generation. Revoke its new
-        # reservation and remove the promoted candidate before surfacing the
-        # failure.
-        try:
-          replace_publication_record(settings, record, "revoked")
-        except (OSError, InvalidPublicationRegistry,
-                PublicationReservationConflict):
-          pass
-      if promoted and record.state == "staged":
-        try:
-          await asyncio.to_thread(shutil.rmtree, destination)
-        except OSError:
-          pass
-      elif promoted and republishing:
-        try:
-          if had_destination:
-            # RENAME_EXCHANGE left the prior complete generation in ``stage``.
-            # Exchange it back before the request reports the metadata failure.
-            atomic_promote_directory(stage, destination)
-          else:
-            # An active record with a missing snapshot was already a 404. Put
-            # that prior state back rather than making failed content public.
-            _rmtree_strict(destination)
-        except OSError as rollback_exc:
-          # The stage is now the only known copy of the previous generation.
-          # Never let the finally block destroy the owner's recovery copy.
-          preserve_stage = True
-          log.error(
-            "republish rollback failed for token %s; prior generation kept "
-            "at %s: %s",
-            token, stage, rollback_exc,
-          )
-          raise rollback_exc from publish_exc
-      raise
-    finally:
-      if not preserve_stage and stage.exists() and not stage.is_symlink():
-        try:
-          await asyncio.to_thread(shutil.rmtree, stage)
-        except OSError:
-          pass
-  return {
-    "token": token,
-    "url": f"/sites/{token}/",
-    "public_url": _publication_url(settings, token),
-  }
-
-
-@router.delete("/{app_id}/publish", dependencies=[Depends(reject_cross_site)])
-async def unpublish_app_site(
-  app_id: int,
-  project_id: str | None = None,
-  db: Session = Depends(get_db),
-  principal: Principal = Depends(get_principal),
-):
-  """Revoke a published URL permanently, then remove its snapshot and hint."""
-  if principal.app_id is not None and principal.app_id != app_id:
-    raise HTTPException(403, "An app may only unpublish its own site.")
-  if project_id and not _PUBLISH_PROJECT_RE.match(project_id):
-    raise HTTPException(422, "invalid project_id")
-  app = live_app_or_404(db, app_id)
-  expected_nonce = app.token_nonce
-  project_id = project_id or None
-  settings = get_settings()
-  async with fs_locks.app_storage_lock(app_id):
-    _recheck_app_identity(db, app_id, expected_nonce)
-    # The registry lives OUTSIDE app-writable storage and is the authority for
-    # what is public, so enumerate + revoke registry-owned tokens first and let
-    # nothing app-controlled gate it. The legacy publish-token.txt hint lives in
-    # app storage, where an app job could leave build/site or the hint itself a
-    # symlink; validating those paths must not be able to keep a registered URL
-    # alive while unpublish 400s.
-    records = [
-      record for record in _registry_records_for_app(settings, app_id)
-      if record.app_gen == app.token_nonce and record.project_id == project_id
-    ]
-    registry_tokens = {record.token for record in records}
-    revoked = True
-    # Revoke registry-owned tokens FIRST and let nothing app-controlled run
-    # before this loop — not even reading the hint. A filesystem error from the
-    # app-writable publish paths (symlink loop, EIO) must never abort unpublish
-    # before the registered URL is dead.
-    for token in registry_tokens:
-      if not await _revoke_publish_token(
-        settings, app_id, app.token_nonce, token, project_id,
-      ):
-        revoked = False
-    # The legacy publish-token.txt hint lives in app storage; read it only if
-    # its paths are sane, and treat ANY error (HTTP validation or raw OS) as
-    # "no legacy hint" rather than letting it block the revoke above.
-    token_file = None
-    hint = None
-    try:
-      _validate_publish_paths(settings, app, project_id)
-      _site, token_file = _publish_paths(settings, app, project_id)
-      hint = _read_publish_token_hint(token_file)
-    except (HTTPException, OSError) as exc:
-      log.warning(
-        "app %s unpublish: skipping legacy hint, publish paths unusable: %s",
-        app_id, exc,
-      )
-      token_file = None
-      hint = None
-    if hint is not None and hint not in registry_tokens:
-      if not await _revoke_publish_token(
-        settings, app_id, app.token_nonce, hint, project_id,
-      ):
-        revoked = False
-    # Only drop the hint once the URL is really dead. Deleting it after a
-    # failed revocation would remove the last pointer to a page that is still
-    # public, and answering {"ok": true} would tell the owner their artifact
-    # was unshared while anyone holding the link could still read it.
-    if not revoked:
-      raise HTTPException(
-        500,
-        "Could not revoke the public URL — it is still live. "
-        "Check storage health and try again.",
-      )
-    if token_file is not None and not token_file.is_symlink():
-      try:
-        token_file.unlink()
-      except OSError:
-        pass
-  return {"ok": True}
+# Compose independently-owned route groups without changing their public paths.
+router.include_router(runtime_router)
+router.include_router(publication_router)
