+++ backend/tests/test_storage.py
@@ -738,12 +738,12 @@ def test_shared_list_paginates_and_skips_symlinks(client, auth):
 def test_numeric_slug_is_prefixed():
   """A purely-numeric name can't become a bare-integer source dir that
   would collide with the numeric-id storage tree (Codex review #4)."""
-  from app.routes.apps import _slugify_for_source_dir
-  assert _slugify_for_source_dir("123") == "app-123"
-  assert not _slugify_for_source_dir("42").isdigit()
+  from app.app_identity import slugify_for_source_dir
+  assert slugify_for_source_dir("123") == "app-123"
+  assert not slugify_for_source_dir("42").isdigit()
   # Non-numeric names are untouched.
-  assert _slugify_for_source_dir("news") == "news"
-  assert _slugify_for_source_dir("Snake 2") == "snake-2"
+  assert slugify_for_source_dir("news") == "news"
+  assert slugify_for_source_dir("Snake 2") == "snake-2"
 
 
 def test_apply_app_rejects_unsafe_source_dir(client, owner_token):
